// angularproject/src/app/matvalidations/matvalidations/matvalidations.interface.ts
export class Matvakidationsinterface {

}
export interface Profilesettinginterface {
    currentemail: string;
    newemail: string;
    confirmemail: string;
}
export interface Profilesettingnumber {
    currentmobilenum: string;
    countrycode: string;
    confmnewph: string;
}
export interface Profilesettingpassword {
    Password: string;
    unewpass: string;
    uconfmnewpass: string;
}
export interface Profilesettingdelete {
    delete: string;
    chkdelete: boolean;
}
// 28_02_2018 Manage Alerts
export interface Managealerts {
    mailyes: string;
    smsyes: string;
}